+++ testsAT/src/test/java/com/stratio/cassandra/lucene/testsAT/deletion/ComposedKeyDataDeletionAT.java
@@ -78,7 +78,7 @@ public class ComposedKeyDataDeletionAT extends BaseAT {
         List<Row> rows = cassandraUtils.delete("bigint_1")
                                        .where("integer_1", 1)
                                        .and("ascii_1", "ascii")
-                                       .refresh()
+                                       .waitForIndexing()
                                        .filter(wildcard("ascii_1", "*"))
                                        .get();
 
@@ -103,7 +103,7 @@ public class ComposedKeyDataDeletionAT extends BaseAT {
         List<Row> rows = cassandraUtils.delete("map_1['k1']")
                                        .where("integer_1", 1)
                                        .and("ascii_1", "ascii")
-                                       .refresh()
+                                        .waitForIndexing()
                                        .filter(wildcard("ascii_1", "*"))
                                        .get();
 
@@ -132,7 +132,7 @@ public class ComposedKeyDataDeletionAT extends BaseAT {
         List<Row> rows = cassandraUtils.delete("list_1[0]")
                                        .where("integer_1", 1)
                                        .and("ascii_1", "ascii")
-                                       .refresh()
+                                       .waitForIndexing()
                                        .filter(wildcard("ascii_1", "*"))
                                        .get();
 
@@ -155,15 +155,11 @@ public class ComposedKeyDataDeletionAT extends BaseAT {
 
     @Test
     public void totalPartitionDeletion() {
-
-        List<Row> rows = cassandraUtils.delete()
-                                       .where("integer_1", 1)
-                                       .and("ascii_1", "ascii")
-                                       .refresh()
-                                       .filter(wildcard("ascii_1", "*"))
-                                       .get();
-
-        assertEquals("Expected 9 results!", 9, rows.size());
-
+        cassandraUtils.delete()
+                      .where("integer_1", 1)
+                      .and("ascii_1", "ascii")
+                      .waitForIndexing()
+                      .filter(wildcard("ascii_1", "*"))
+                      .check(9);
     }
 }