+++ testsAT/src/test/java/com/stratio/cassandra/lucene/testsAT/util/CassandraConfig.java
@@ -29,7 +29,7 @@ public class CassandraConfig {
     static final int FETCH = Integer.parseInt(get("fetch", "100"));
     static final int THREADS = Integer.parseInt(get("threads", "0"));
     static final int REFRESH = Integer.parseInt(get("refresh", "1"));
-    static final int WAIT_FOR_INDEXING = Integer.parseInt(get("wait_for_indexing", "1"));
+    static final int WAIT_FOR_INDEXING = Integer.parseInt(get("wait_for_indexing", "2"));
     static final int WAIT_FOR_SERVER = Integer.parseInt(get("wait_for_server", "30"));
     static final String TABLE = get("table", "test_table");
     static final String INDEX = get("index", "test_table_idx");