+++ testsAT/src/test/java/com/stratio/cassandra/lucene/testsAT/util/CassandraUtilsUpdate.java
@@ -17,6 +17,8 @@ package com.stratio.cassandra.lucene.testsAT.util;
 
 import com.datastax.driver.core.querybuilder.QueryBuilder;
 import com.datastax.driver.core.querybuilder.Update;
+import com.stratio.cassandra.lucene.builder.search.condition.Condition;
+import com.stratio.cassandra.lucene.builder.search.sort.SortField;
 
 /**
  * @author Andres de la Pena {@literal <adelapena@stratio.com>}
@@ -45,10 +47,27 @@ public class CassandraUtilsUpdate {
         update.where().and(QueryBuilder.eq(name, value));
         return this;
     }
-
-    public CassandraUtils refresh() {
+    private CassandraUtils execute() {
         parent.execute(update);
-        parent.refresh();
         return parent;
+    }
+    public CassandraUtils refresh() {
+        return execute().refresh();
+    }
+
+    public CassandraUtilsSelect query(Condition query) {
+        return execute().query(query);
+    }
+
+    public CassandraUtilsSelect filter(Condition filter) {
+        return execute().filter(filter);
+    }
+
+    public CassandraUtilsSelect sort(SortField... sort) {
+        return execute().sort(sort);
+    }
+
+    public CassandraUtils waitForIndexing() {
+        return execute().waitForIndexing();
     }
 }